déplacement des fonctions de requêtes cocktail de crud_db vers cocktail_db
