+++ back-end/flask_app/api/crud_db.py
@@ -37,6 +37,7 @@ def read_data():
     content = request.json
     view_query = []
     pulled_data = coll.find({"name":content['name']})
+    # pulled_data = coll.find({}, {'_id':False})
     for e in pulled_data:
         view_query.append({"name":e['name'], "alcohol": e['alcohol']})
     return jsonify(view_query)
@@ -53,47 +54,3 @@ def delete_data():
     content = request.json
     coll.delete_one({"name":content['name']})
     return "Data deleted"
-
-
-# Créer un cocktail
-def create_cocktail():
-    data = request.json
-    try:
-        new_cocktail = {
-            "Nom": data["Nom"],
-            "Alcool": data["Alcool"],
-            "Degres": data["Degres"],
-            "Ingredients": data["Ingredients"],
-            "Recette": data["Recette"],
-            "Auteur": data["Auteur"],
-            "Date_Creation": datetime.utcnow()
-        }
-        coll.insert_one(new_cocktail)
-        return jsonify({"message": "Cocktail ajouté !"}), 201
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-# Lecture de  tous les cocktails
-def get_cocktails():
-    cocktails = list(coll.find({}, {"_id": False}))
-    return jsonify(cocktails)
-
-# Lecture d'un cocktail par son nom
-def get_cocktail_by_name():
-    data = request.json
-    cocktail = coll.find_one({"Nom": data["Nom"]}, {"_id": False})
-    if cocktail:
-        return jsonify(cocktail)
-    return jsonify({"message": "Cocktail introuvable"}), 404
-
-# Maj d'un cocktail
-def update_cocktail():
-    data = request.json
-    coll.update_one({"Nom": data["Nom"]}, {"$set": data})
-    return jsonify({"message": "Cocktail mis à jour !"})
-
-# Supprimer un cocktail
-def delete_cocktail():
-    data = request.json
-    coll.delete_one({"Nom": data["Nom"]})
-    return jsonify({"message": "Cocktail supprimé !"})
